+++ src/components/banner/no-match-banner.tsx
@@ -2,17 +2,20 @@ import { useRecoilState } from 'recoil';
 import BannerIcon from 'components/banner/item/banner-icon';
 import CancelButtonXs from 'components/button/cancel-button-xs';
 import YellowButtonXs from 'components/button/yellow-button-xs';
+import BannerMessage from 'components/banner/item/banner-message';
 import { GameBanner } from 'ts/enums/game/game-banner.enum';
 import { gameBannerState } from 'ts/states/game/game-banner-state';
-import { NO_MATCH_MSG, RE_WAIT_MSG } from 'constants/messages';
-import BannerMessage from 'components/banner/item/banner-message';
+import { gameWaitState } from 'ts/states/game/game-wait-state';
+import { GameWaitStatus } from 'ts/enums/game/game-wait.enum';
+import { NO_MATCH_MSG } from 'constants/messages';
 
 const NoMatchBanner = () => {
 	const [gameBanner, setGameBanner] = useRecoilState(gameBannerState);
+	const [gameWait, setGameWait] = useRecoilState(gameWaitState);
 
 	const handleReMatch = () => {
 		setGameBanner({ ...gameBanner, banner: GameBanner.NONE });
-		// 모달 띄우기
+		setGameWait({ ...gameWait, status: GameWaitStatus.SEARCH });
 	};
 
 	const handleCancel = () => {
@@ -30,7 +33,7 @@ const NoMatchBanner = () => {
 						</div>
 						<div className='flex items-center'>
 							<YellowButtonXs onClick={handleReMatch}>
-								{RE_WAIT_MSG}
+								다시 기다리기
 							</YellowButtonXs>
 							<CancelButtonXs onClick={handleCancel} />
 						</div>
