+++ src/components/modal/game-search-modal.tsx
@@ -2,27 +2,26 @@ import { useState } from 'react';
 import { useRecoilState } from 'recoil';
 import { Modal } from 'flowbite-react';
 import UseCountdown from 'hooks/use-countdown';
-import { GameBanner } from 'ts/enums/game/game-banner.enum';
-import { gameBannerState } from 'ts/states/game/game-banner-state';
 import WhiteButtonXl from 'components/button/white-button-xl';
 import TimeSpinner from 'components/modal/item/time-spinner';
+import { GameBanner } from 'ts/enums/game/game-banner.enum';
+import { gameBannerState } from 'ts/states/game/game-banner-state';
+import { gameWaitState } from 'ts/states/game/game-wait-state';
+import { GameWaitStatus } from 'ts/enums/game/game-wait.enum';
+import ModalProps from 'ts/interfaces/game/modal-props';
 import {
 	COUNTDOWN_MATCH_INTERVAL,
 	COUNTDOWN_MATCH_VALUE,
 } from 'constants/values';
 
-interface ModalProps {
-	show: boolean;
-	onClose: () => void;
-}
-
 const GameSearchModal: React.FC<ModalProps> = ({ show, onClose }) => {
 	const [seconds, setSeconds] = useState(COUNTDOWN_MATCH_VALUE);
+	const [gameWait, setGameWait] = useRecoilState(gameWaitState);
 	const [gameBanner, setGameBanner] = useRecoilState(gameBannerState);
 
 	const handleNoMatch = () => {
 		setGameBanner({ ...gameBanner, banner: GameBanner.NOMATCH });
-		onClose();
+		setGameWait({ ...gameWait, status: GameWaitStatus.NONE });
 	};
 
 	const handleCancel = () => {
@@ -30,7 +29,7 @@ const GameSearchModal: React.FC<ModalProps> = ({ show, onClose }) => {
 	};
 
 	const handleMatch = () => {
-		// 매칭 이을 곳
+		// 소켓 이을 곳
 		onClose();
 	};
