+++ src/components/modal/game-wait-modal.tsx
@@ -1,13 +1,9 @@
 import { useRecoilValue } from 'recoil';
 import GameMatchModal from 'components/modal/game-search-modal';
 import GameModeModal from 'components/modal/game-mode-modal';
+import ModalProps from 'ts/interfaces/game/modal-props';
 import { GameWaitStatus } from 'ts/enums/game/game-wait.enum';
 import { gameWaitState } from 'ts/states/game/game-wait-state';
-
-interface ModalProps {
-	show: boolean;
-	onClose: () => void;
-}
 
 const GameWaitModal: React.FC<ModalProps> = ({ show, onClose }) => {
 	const gameWait = useRecoilValue(gameWaitState);
