+++ src/constants/messages.ts
@@ -7,9 +7,3 @@ export const NO_MATCH_MSG =
 	'경기를 원하는 다른 사용자가 없어서 게임이 취소되었습니다.';
 
 export const GAME_REQ_MSG = '님이 게임 요청을 보냈어요!';
-
-export const RE_WAIT_MSG = '다시 기다리기';
-
-export const GAME_APPROVE_MSG = '승낙하기';
-
-/* game - modal */
