+++ src/constants/values.ts
@@ -2,5 +2,5 @@
 export const COUNTDOWN_REQUEST_VALUE: number = 100;
 export const COUNTDOWN_REQUEST_INTERVAL: number = 20;
 
-export const COUNTDOWN_MATCH_VALUE: number = 10;
+export const COUNTDOWN_MATCH_VALUE: number = 2;
 export const COUNTDOWN_MATCH_INTERVAL: number = 1000;
