--- src/ts/interfaces/game/game-modal-control.ts
@@ -1,9 +1,0 @@
-import { GameMode } from 'ts/enums/game/game-mode.enum';
-
-interface GameOptionControl {
-	isOpen: boolean;
-	onClose: () => void;
-	onMatch: (mode: GameMode) => void;
-}
-
-export default GameOptionControl;
