+++ src/ts/interfaces/game/modal-props.ts
@@ -1,0 +1,6 @@
+interface ModalProps {
+	show: boolean;
+	onClose: () => void;
+}
+
+export default ModalProps;
